pyproject.toml
[tool.poetry]
name = "news-site-backend"
version = "0.1.0"
description = ""
authors = ["Anastasia Sypkova <a.sypkova@tg.dunice.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
django-cors-headers = "^4.3.1"
djangorestframework-camel-case = "^1.4.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.0.0"
pre-commit = "^3.7.0"
flake8-bugbear = "^24.4.26"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
